requirements.txt
python-dotenv
google-genai>=1.68.0
tenacity>=8.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langfuse>=3.0.0